correction d'un bug

+++ cant_stop/environments/gym_env_v3.py
@@ -127,7 +127,8 @@ class CantStopGymEnv(gym.Env):
                 print(f"Reward : {self.reward}")
                 self.render()
                 sys.exit()
-            self.reward = - (self.reward**(3/2))
+            self.reward = int(- (self.reward**(3/2)))
+            
             # print(f"RL a busté, reward : {self.reward}")
             self.turn += 1
             self.reward += 0
